Simplify utils back to a single file and fix any build issues.
