+++ workspace/all/clock/clock.c
@@ -6,7 +6,6 @@
 #include <unistd.h>
 
 #include "api.h"
-#include "date_utils.h"
 #include "defines.h"
 #include "utils.h"
