--- workspace/all/common/utils/utils.h
@@ -1,9 +1,0 @@
-#ifndef UTILS_H
-#define UTILS_H
-
-#include <stdint.h>
-
-// Timing functions
-uint64_t getMicroseconds(void);
-
-#endif
